Detect Wad Type: Support THUG 2 Format
-I have updated the about page to include THUG2 for listed supported
files as THUG2 PSP uses the same .hed/.wad format as THP8's
musicp/streamsp .hed/.wad files.

-Previously Thug 2 .hed/.wad files could be extracted, but only though
the musicp/streamsp menu. This creates confusion when using the THUG2
datap file, which is the same format as the musicp/streamsp.

Now the program detects if which format it is, so that you can open the
THUG2 datap via the datap menu successfully. Though it will extract
correctly, currently you can not rebuild it via the datap menu.

Once this format is supported for rebuilding it will detect on creation.
Though, it should not create an error if one attempts to build it via
the datap menu, but it may not function in game.

-Rearranged certain functions in the source file for readability.

-Updated .EXE to match current build.